requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
werkzeug==3.0.1
pandas~=2.0.3
matplotlib==3.8.2
ta==0.11.0
requests==2.31.0
groq==0.4.2
python-dotenv==1.0.0
gunicorn==21.2.0
MetaTrader5==5.0.45
cryptography==41.0.7
schedule==1.2.1
email-validator==2.1.0.post1